redesign candidate profile with tabs, id display, and quick reject
- Add C001 format display_id to candidate API
- Refactor profile page: header + 3 tabs (resume/pipeline/history)
- Header shows name@id, latest work/edu, contact, current pipeline status
- Pipeline tab supports job transfer and reject with interview record
- Pipeline tracking table adds quick reject button with interview prompt
- years_exp changed to Float for 0.5 precision
- AI prompt updated for 0.5-year work experience precision

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, DateTime, ForeignKey, JSON
+from sqlalchemy import Column, Integer, String, Text, DateTime, ForeignKey, JSON, Float
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from app.database import Base
@@ -17,7 +17,7 @@ class Candidate(Base):
     city = Column(String)
     last_company = Column(String)
     last_title = Column(String)
-    years_exp = Column(Integer)
+    years_exp = Column(Float)
     name_en = Column(String)
     education_list = Column(JSON, default=list)   # [{degree, school, major, period}]
     work_experience = Column(JSON, default=list)  # [{company, title, period}]
